Implemented baseline run and tuned tqdm output
