+++ src/reports/adoc/regression.py
@@ -6,7 +6,8 @@ from sql_formatter.core import format_sql
 from matplotlib import pyplot as plt
 from matplotlib import rcParams
 
-from objects import CollectResult, Query
+from collect_objects import CollectResult
+from db_objects import Query
 from reports.abstract import Report
