+++ src/reports/adoc/score.py
@@ -5,8 +5,9 @@ from matplotlib import pyplot as plt
 from matplotlib import rcParams
 from sql_formatter.core import format_sql
 
+from collect_objects import CollectResult
 from db.postgres import PostgresQuery
-from objects import CollectResult, Query
+from db_objects import Query
 from reports.abstract import Report
 from utils import allowed_diff
