+++ src/reports/adoc/taqo.py
@@ -1,10 +1,10 @@
-import os
 from typing import Type
 
 from matplotlib import pyplot as plt
 from sql_formatter.core import format_sql
 
-from objects import CollectResult, Query
+from collect_objects import CollectResult
+from db_objects import Query
 from reports.abstract import Report
 from utils import allowed_diff
