+++ src/utils.py
@@ -10,7 +10,7 @@ import psycopg2
 
 from config import Config
 from db.database import Database
-from objects import Query, FieldInTableHelper
+from db_objects import Query, FieldInTableHelper
 
 PARAMETER_VARIABLE = r"[^'](\%\((.*?)\))"
 WITH_ORDINALITY = r"[Ww][Ii][Tt][Hh]\s*[Oo][Rr][Dd][Ii][Nn][Aa][Ll][Ii][Tt][yY]\s*[Aa][Ss]\s*.*(.*)"
